Remove window draw radius and fix framed mouse capture area on windows
1px draw radius is unnecessary for nw and causes a 1px artifact at the corner of all windows
Fixes rogerwang/node-webkit#331

We create a mouse dead zone on frameless windows, because the drag handles are only available within the window border -- not outside (as the case is for framed windows). The if block was incorrectly only covering the dead zone on one side of the window -- not all four.
Fixes rogerwang/node-webkit#321