Use the window title when creating OS X menu bar entries.
This implements a chain of possible names for the OS X menu
entries. Previously this was always the 'name' field, which is a logical
name. This changes it so it defaults to the main window's title; if that
is not set, it falls back to the name in the manifest.

If the name is 'node-webkit' (the default or unset), it uses the value
of CFBundleName.

This problem is described in issue #1847.