[docs] update for Multiple Context mode
The document confuses user on this topic. There is nothing prevent user
from writing memory leaking code either when you are using Node or not,
even in browser. Especially when you're writing long live single page
applications/pages. It's a bigger topic not specific to NW.

The mixed context mode is created to provide isolation to use multiple
Node contexts, not a solution to memory management.