Fix of not finding menubar button when navigate to another menu
When pressing a menubar button and navigate to another, such as
clicking `File` and navigate to `Edit`, `MenuModelAdapter` will
still try to find buttons within the model which it was created
with, which is `File` in this case. The fix is to record current
active model and override two `ExecuteCommand` methods to find
buttons in the correct model.

Fixed #4371