implemented `win.menu=null` to remove menubar
Previously in 0.12.x the visibility of menubar on fullscreen/kiosk
 mode is inconsistent on different platforms. Since 0.13.0, the
menubar is always visible on fullscreen / kiok mode. Using
`win.menu=null` can remove the menubar manually. This gives
developers better control on the menubars.

This patch implemented `win.menu=null` to completely remove the
menubar on Linux and Windows, and cleared out the menubar on Mac.
The patch also updated the docs about menubar visibility on
fullscreen mode.

Fixed #4725