Update instructions
I ran into a couple of issues trying to build the project for the
firt time, here are some updates to the documentation that would have
saved me many hours.

In particular I didn't realize that I didn't want to build nw17, and
that even after updating that to nw25, I also had to check out the
correct version of every dependency.

I also didn't know how to set build arguments (had to Google that),
and ran into the problem with gn being missing, so I copied the
instructions from the "Enable Proprietary Codecs" page.