Incorrect guidelines for selenium test automation
I tried exactly what was given in the current version of documentation for test automation using selenium on the nw.js application. According to the documentation, copying only the chrome driver from the SDK build to the application directory should work. But it didn't work for me and was throwing an error saying chrome not reachable. Fortunately, it worked when I built the application with the SDK version of nw.js. So I made a correction in the documentation on the same.