Update Building NW.js.md
This is to ensure the documentation is update to
support latest branch to avoid branch confusions
among all the repos required in building nw.js.

Since chromium has started to shift focus to GN,
GYP variables are now required explicit export
in order to be able to generate ninja build
files with GYP for Node.

Also, the gyp_chromium.py was written in Python2
syntax, hence any host that has Python3 set as
default will have issue running the command.

Signed-off-by: Rebecca Chang Swee Fun <rebecca.chang@starfivetech.com>